Final cleanup: extract shared ASR utils, fix review comments
DRY refactor:
- Extract shared audio preprocessing into examples/asr_utils.py:
  load_audio_file, compute_fbank, apply_lfr, load_cmvn, preprocess_audio
- Both fun_asr.py and sensevoice_small.py now import from asr_utils

Review comment fixes:
- Fix stale docstring: seq_len//2 → seq_len in FunASRAudioEncoder
- Fix AudioCTCTask: use config.dtype instead of hardcoded FLOAT
- Add explanatory comments to all empty except clauses
